add surface to the phoenix liveview application
Before setup, the surface throughout the project, was necessary to add surface as a dependency, to do that, it needed to follow these steps:
- add the `surface` `0.6.x` version
- should be necessary to downgrade the `phoenix_live_view` to `0.16.x`
  version.
- should be necessary to downgrade the `phoenix_live_dashboard` to `0.5.x`
  version.

After that get the dependency and compile it.
So now, it could be run as a mix task.
And this mix task accepts options like `--demo` and `--catalogue`.

- `mix surface.init`
- `mix surface.init --demo --catalogue`

So to get more information about it, it could run this mix task with
help. Like: `mix help surface.init`.

Also, this PR was a runner with `--demo` opts.

```shell
$ mix surface.init --demo
```

After that, it is possible to start the application and browser to [http://localhost:4000/demo](http://localhost:4000/demo) to see the first example with `surface`.

Reference:
- https://surface-ui.org/getting_started