Use production config from Heroku ENV variables and decrease socket timeout (#8)
fix always heroku rebuild

This has to do with stale dependencies which are not getting recompiled properly. It's possible to force Heroku to recompile all dependencies on each deploy, which should fix this problem. The way to do it is to add a new file called elixir_buildpack.config at the root of the application. The file should contain this line:

```
always_rebuild=true
```

Commit this file to the repository and try to push again to Heroku.

bump elixir and erlang/otp version

remove the websocket timeout

prepares for heroku

more prepares for heroku

fix prepares for heroku